requirements.txt
Flask==0.12.2
grpcio>=1.4.0, <1.6.0
google-cloud==0.27.0
gunicorn==19.7.1
mock==2.0.0
Flask-SQLAlchemy==2.3.0
PyMySQL==0.7.11
Flask-PyMongo==0.5.1
PyMongo==3.5.1
six==1.11.0
requests[security]==2.18.4
honcho==1.0.1
psq==0.5.0
datadog
ddtrace
